+++ cxbin/loaderimpl.h
@@ -25,8 +25,8 @@ namespace cxbin
 		virtual ~LoaderImpl() {}
 
 		virtual std::string expectExtension() = 0;
-		virtual bool tryLoad(FILE* f, unsigned long long fileSize) = 0;
-		virtual bool load(FILE* f, unsigned fileSize, std::vector<trimesh::TriMesh*>& out, ccglobal::Tracer* tracer) = 0;
+		virtual bool tryLoad(FILE* f, size_t fileSize) = 0;
+		virtual bool load(FILE* f, size_t fileSize, std::vector<trimesh::TriMesh*>& out, ccglobal::Tracer* tracer) = 0;
 	};
 
 	CXBIN_API void registerLoaderImpl(LoaderImpl* impl);
